Allow out-of-tree builds
When building out of tree, the value of ctypes_la_CPPFLAGS =
-I../include is wrong; it points to a location in the build tree, not
the source tree as intended.

See
https://www.gnu.org/software/automake/manual/html_node/Program-Variables.html,
specifically the notes regarding $(srcdir) vs $(builddir) and whey they
matter for out of tree builds.

This commit replaces the hardcoded path ../include and ../libs with
the calculated equivalents $(srcdir)/{include,libs} - this allows
building the package in-tree or out-of-tree