Replace the configure script with INSTALL instructions
The configure script should not be committed to source control; it is a
generated file.

Instead, instructions on how to bootstrap the package when checked out
of a source control system are given in the INSTALL file.

This commit removes the configure script and adds the needed INSTALL
file.

The motivation is to make the package consistent with autotools best
practices and expectations of users who are familiar with autotools.

Secondarily, the standard bootstrap-configure-make workflow is well
known, and is implemented in the VSCode Autotools Integration extension,
which I wrote. The extension is able to handle the well-known package
lifecycle without user intervention.

The unusual presence of a 'configure' script which is not a
configuration script confused my extension, as it would confuse any
automated system. The extension would run the configure script,
expecting that it would create the Makefiles, but it had to stop
processing when the Makefiles are not created, promting the user to
figure out what to do next.

This change implements the well known workflow, allowing the package to
be handled in a completely automated manner where needed, while
providing users who are initiated in building software with autotools
the familiar interface 'bootstrap/autogen, configure, make'